Устранен баг заргузки/отображения картинок

+++ src/main/java/ru/geekbrains/habr/services/ArticleService.java
@@ -118,14 +118,14 @@ public class ArticleService {
         if (article.getTitle().equals(articleDto.getTitle())
                 && article.getText().equals(articleDto.getText())
                 && article.getStatus().equals(articleDto.getStatus())
-                && article.getImagePath()!=null && !article.getImagePath().equals(articleDto.getImagePath()))
+                && article.getImagePath().equals(articleDto.getImagePath()))
             return;
 
         article.setText(articleDto.getText());
         article.setTitle(articleDto.getTitle());
         article.setStatus(articleDto.getStatus());
 
-        if(article.getImagePath()!=null){
+        if(article.getImagePath()!=null && !article.getImagePath().equals(articleDto.getImagePath())){
             imageService.deleteImage(article.getImagePath());
         }
         article.setImagePath(articleDto.getImagePath());
